requirements/base.txt
python-dotenv==1.0.1
spacy==3.7.5
nltk==3.8.1
transformers==4.43.3
openai==0.28.0
tensorflow==2.17.0
tf_keras==2.17.0
pandas==2.2.2
tiktoken==0.7.0
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.22
youtube-search==2.1.2
requests==2.32.3
beautifulsoup4==4.12.3